requirements.txt
streamlit>=1.30.0
python-docx>=1.1.0
lxml>=4.9.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
requests>=2.31.0
bibtexparser>=1.4.0
